[nemo-qml-plugin-calendar] Specify the notebook when purging.
Provides the notebook UID of the event to
purge. Not a necessity at the moment since
event UIDs are unique in the database, but
it will allow to raise the unicity constrain
between notebooks later on.

+++ src/calendarworker.cpp
@@ -130,7 +130,7 @@ void CalendarWorker::storageUpdated(mKCal::ExtendedStorage *storage,
         // if the event was stored in a local (non-synced) notebook, purge it.
         const CalendarData::Notebook &notebook = mNotebooks.value(mCalendar->notebook(event));
         if (notebook.localCalendar
-            && !storage->purgeDeletedIncidences(KCalendarCore::Incidence::List() << event)) {
+            && !storage->purgeDeletedIncidences(KCalendarCore::Incidence::List() << event, notebook.uid)) {
             qWarning() << "Failed to purge deleted event" << event->uid()
                        << "from local calendar" << mCalendar->notebook(event);
         }
